requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Cors==4.0.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
Werkzeug==3.0.3
pytest==8.2.2
pytest-flask==1.3.0
